Renamed files to get them excluded from serverless functions build
