+++ app/api/raffle/route.ts
@@ -2,8 +2,8 @@ import { prisma } from "@/lib/prisma";
 import { getServerSession } from "next-auth";
 import { NextRequest, NextResponse } from "next/server";
 import { authOptions } from "../auth/[...nextauth]/route";
-import { createRewardsSub, deleteListener } from "../webhooks/helpers";
-import { AggregateGiveaway } from "../statistics/helpers";
+import { createRewardsSub, deleteListener } from "../webhooks/_helpers";
+import { AggregateGiveaway } from "../statistics/_helpers";
 
 /** Create a new reward and save it to the db */
 export async function PUT(req: Request) {
