+++ app/api/statistics/_helpers.ts
